Ensure more strictly that Jenkins is up
We had problems that looked like a possible race condition
when adding more plugins with the current install. Restarting
jenkins seemed to take too much time. The plain wait_for for
the socket to open did not solve the whole issue, but then the
/cli/ URL was not responding.

The pull request content is not ideal by any means (the curl
check to see that Jenkins is fully started is triplicated),
but maybe you get the idea.