Add jenkins_cli_extra_opts variable.
In some cases, one may need to use an extra CLI option, in particular,
`-noKeyAuth`, or perhaps the `-i ssh_key` option. Adding this variable,
which defaults to the empty string, one can add CLI options on the
ansible-playbook command line, for example.

A particular situation is when running the CLI within Ansible as the
Jenkins user - is there really a point in using SSH credentials when
you are already the Jenkins user?